Add Espressif's HMI DevKit
this is only tested to come up to the REPL & mount CIRCUITPY. Pin
assignments should be right but were not double-checked.  The screen
is unsupported so far.

This board depends on the I/O pull ups for the I2C bus (verified by
schematic) so this adds a compile time option that enables pull ups
for ANY i2c bus on a board.