Selectively build boards based on changed files
This uses a step output from the test job to set the build matrix
for the board build jobs. The built boards depends on which files
were changed.

* Changes contained within ports/*/boards/ will build each board.
* Changes contained within ports/* will build all boards from the
  port.
* All other file changes will build all boards.

All boards will be build for pushes in `adafruit/circuitpython` as
well.

A side-effect is that we no longer need to explicitly list the
boards to build. It will automatically update as new folders are
added.

Related to #4009