ports/stm32/boards: Add OLIMEX H407 board support.
This change adds the OLIMEX H407 support to the STM32 port.
The H407 (https://www.olimex.com/Products/ARM/ST/STM32-H407/) is simliar
to the already existing E407
(https://www.olimex.com/Products/ARM/ST/STM32-E407) but does not support
Ethernet and has a full-size USB-A port instead of a Mini-USB socket.

Both boards use the STM32F407ZGT6 CPU.

This port is basically a copy of the E407 but with changed pinmux:
* Removed Ethernet pin definition
* Removed UART1 (pins are used for other functions)
* Removed UART3 flow control pins (pins are used for other functions)
* Removed SD-Card detect pin (since it is not connected on the H407)

I have added a REPL on UART3 since this is connected to the
U3BOOT-header. This is a 3-pin header with RX, TX and GND that is
inteded for the serial terminal.

Tested:
* Micro-SD Card is detected when inserted on RESET
* REPL on UART3 works
* Serial port on the mini USB socket

Signed-off-by: Chris Fiege <cfi@pengutronix.de>