mimxrt1170: Add the 1176 port in sync with MP Master.
Even if the RT1176 has two cores, the actual firmware supports only the
CM7. We have no good plan how to use the CM4 (see also the discussion
about the Portenta H7). The actual MIMXRT1176_EVK port is on par
with the existing MIMXRT boards, with the following extensions:

- use 64 MB RAM for the heap.
- support both LAN interfaces as LAN(0) and LAN(1), with LAN(1)
being the 1GB interface.

The dual LAN port interface can be adapted as well for the RT1062 mcu.
Only I have no 1062 board for testing with 2 LAN ports.