py/runtime: Adjust message for not defined name
For mainstream Python, the message is `name '...' is not defined`:

```
$ python3
Python 3.8.10 (default, Sep 28 2021, 16:10:42)
[GCC 9.3.0] on linux
Type "help", "copyright", "credits" or "license" for more information.
>>> forsuredoesnotexist
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
NameError: name 'forsuredoesnotexist' is not defined
>>>
```

I found this discrepancy when working on a project that uses Python
under the hood, and tried to replace regular Python with MicroPython.
The discrepancy was shown as a difference in behavior between regular
Python and MicroPython, and then I realized that's it's only about
wording in English. I propose unifying this message, for the sake of
bringing both pythons closer to each other.

In theory there's a risk that someone depends on this string, which
would make it a breaking change. However, I assess the risk as
extremely low, and still the change brings consistency with the regular
Python.