ports/esp32: Fix SPI default pins reordering on ESP32S2 and S3.
The index of machine_hw_spi_obj and machine_hw_spi_default_pins
arrays is assigned to 0 for ARG_id==HSPI_HOST and 1 for another SPI.
On ESP32S2 and S3 HSPI_HOST=2 so the first set (idx=0) of default
pins is used for SPI(id=2) aka HSPI/SPI3 and the second set (idx=1)
for SPI(id=1) aka FSPI/SPI2.
This makes misleading mess in MICROPY_HW_SPIxxxx definitions
and it is also in contradiction to the comments around the definitions.

Change the test of ARG_id to fix the order of machine_hw_spi_default_pins.

This change might require adjusting MICROPY_HW_SPIxxxx definitions
in mpconfigboard.h of S2/S3 based boards.

Signed-off-by: Tomas Vanek <vanekt@fbl.cz>