update cmake support
This fix adds support for CMAKE generator expressions in include paths and definitions.

If _arg in the modified foreach syntax is composed of generator expressions, then
e.g. $<$<config:Debug>:DEBUG=1>.
MICROPY_CPP_FLAGS will contain -D. This will result in an error in gcc.

Therefore, a fix is needed to use generator expressions and not generate it if it is empty.