py/qstr: Reset mpstate.qstr_last_chunk before raising an error.
The qstr_last_chunk is not collected by the garbage collector. This
relies on the assertion that qstr_pool_t also references the
qstr_last_chunk. If an exception is raised while allocating the
qstr_pool_t, qstr_last_chunk has to be invalidated not to become
a dangling reference at the next garbage collection.

Signed-off-by: Emilie Feral <emilie.feral@numworks.com>