mimxrt: Fix an Ethernet transmit error.
Sometimes frames could not be sent immediately because the controller
was still busy with previous frames. Then, an error was returned to
lwip. This fix adds a limited number of retries for this busy state,
waiting 100µs before the next attempt. Typically the transmit succeeds
now at the second attempt.

Second change: Reset the controller for a clean state after soft reset.