rp2: Change the RTS behavior.
When the buffer is full, the UART fifo will not be cleared now. Tested
with sending 1000 bytes and reading in chunks of 100, 200, 300, 400,
500 bytes with 100 ms delay between the read calls.

The sender may time out if the total overal read rate is too slow. But
in any case the number of bytes sent matches the number of bytes
received.