wm8960: Driver for the WM8960 codec.
This codec is assembled for the MIMXRT1xxx_DEV boards and available
for WM8960 breakout boards as well.

The driver itself has been tested as working with the MIMXRT boards
and a Sparkfun WM6890 breakout board. It implements the initialization,
basic methods and some enhanced methods like 3D, ALC or Soft Mute.