tests/run-tests.py: Skip repl tests when running windows underneath.
Some versions of Python (for instance: the mingw-w64 version which
can be installed on MSYS2) do include a pty module and claim to be
posix-like (os.name == 'posix'), yet the select.select call used in
run-tests.py hangs forever.  To be on the safe side just exclude
anything which might be running on windows.