windows: Consolidate all sleep-related functions into windows_mphal.c.
Replace the timer-based sleep with the standard win32 call since the
former has no benefits: even though it allows specifying the time in
100uSec chunks, the actual resolution is still limited by the OS and
is never better than 1mSec.
For clarity move all of this next to the mp_hal_delay_ms definition so
all related functions are in one place.