py/parse.c: Implement function keyword argument syntax restriction.
Implemented function keyword argument syntax restriction added with
Python 3.8 (listed under issue #7899). This is done by adding a switch
in the push_result_rule function in parse.c to turn off the parentheses
removal optimization for atom_paren rules encountered while parsing
an arglist rule (meaning this atom_paren rule is a function argument)
and when the enclosed node is a lone ID. Keeping the atom_paren rule
in the parse tree allows the compiler to catch the syntax error when
the rule happens to be a keyword argument.

In order to determine whether an atom_paren rule is a function argument,
a bit is added to the parser structure, which is set in mp_parse when
an arglist rule is being parsed (unset once done with it). Other
options were considered: hard-coding the distance on the rule stack
between an arglist rule and an atom_paren rule in the added switch,
using a global boolean variable, using a local boolean variable (by
expanding the function signature of the push_result_rule function).
Adding a field to the parser structure seemed to provide the best
balance between code size increase, code changes, memory usage
overhead, code maintainability, etc.

Because the compile_atom_paren function did not expect this rule to
be a parent for a lone ID, a switch is added in that function to
handle this.

A new test file fun_kwargs_syntax.py is added to make sure the
syntax restriction is in effect, and a test is added to fun2.py to
make sure extraneous parenthese around ID function arguments does not
cause any issues.

Signed-off-by: Rayane Chatrieux <rayane.chatrieux@gmail.com>