tool/gen-cpydiff: Skip Black fmt comments.
Since cpydiff is code used as documentation, there are cases where we
may want to use Black's `fmt: on/off/skip` comments to avoid automatic
formatting. However, we don't want these comments to be distracting in
the generated documentation.

This rewrite the code to omit these comments when generating the docs.

Signed-off-by: David Lechner <david@pybricks.com>